// src/app/api/fn/partners/partners-suppliers-list.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PaginatedSupplierList } from '../../models/paginated-supplier-list';

export interface PartnersSuppliersList$Params {

/**
 * true: active only; false: inactive only.
 */
  active?: 'false' | 'true';

/**
 * A page number within the paginated result set.
 */
  page?: number;

/**
 * Matches codes, names and numbers.
 */
  search?: string;

/**
 * Supplier type.
 */
  supplier_type?: 'DISTRIBUTOR' | 'IMPORTER' | 'LOCAL_MARKET' | 'MANUFACTURER' | 'SERVICE_CENTRE' | 'WHOLESALER';
}

export function partnersSuppliersList(http: HttpClient, rootUrl: string, params?: PartnersSuppliersList$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginatedSupplierList>> {
  const rb = new RequestBuilder(rootUrl, partnersSuppliersList.PATH, 'get');
  if (params) {
    rb.query('active', params.active, {});
    rb.query('page', params.page, {});
    rb.query('search', params.search, {});
    rb.query('supplier_type', params.supplier_type, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginatedSupplierList>;
    })
  );
}

partnersSuppliersList.PATH = '/api/partners/suppliers/';